New BinaryData API, make BinaryDataContent a public API
