+++ sdk/core/azure-core-serializer-json-jackson/src/main/java/com/azure/core/serializer/json/jackson/implementation/BinaryDataSerializer.java
@@ -3,15 +3,8 @@
 
 package com.azure.core.serializer.json.jackson.implementation;
 
-import com.azure.core.implementation.util.BinaryDataContent;
-import com.azure.core.implementation.util.BinaryDataHelper;
-import com.azure.core.implementation.util.ByteArrayContent;
-import com.azure.core.implementation.util.FileContent;
-import com.azure.core.implementation.util.FluxByteBufferContent;
-import com.azure.core.implementation.util.InputStreamContent;
-import com.azure.core.implementation.util.SerializableContent;
-import com.azure.core.implementation.util.StringContent;
 import com.azure.core.util.BinaryData;
+import com.azure.core.util.binarydata.BinaryDataContent;
 import com.azure.core.util.logging.ClientLogger;
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.JsonSerializer;
@@ -43,19 +36,23 @@ final class BinaryDataSerializer extends JsonSerializer<BinaryData> {
             return;
         }
 
-        BinaryDataContent content = BinaryDataHelper.getContent(value);
-        if (content instanceof ByteArrayContent
-            || content instanceof FileContent
-            || content instanceof FluxByteBufferContent
-            || content instanceof InputStreamContent) {
-            gen.writeBinary(content.toBytes());
-        } else if (content instanceof SerializableContent) {
-            gen.writeRawValue(content.toString());
-        } else if (content instanceof StringContent) {
-            gen.writeString(content.toString());
-        } else {
-            throw LOGGER.logExceptionAsError(
-                new IllegalStateException("Unsupported BinaryData content type: " + content.getClass().getName()));
+        BinaryDataContent content = value.getContent();
+        switch (content.getContentType()) {
+            case BINARY:
+                gen.writeBinary(content.toBytes());
+                break;
+
+            case OBJECT:
+                gen.writeRawValue(content.toString());
+                break;
+
+            case TEXT:
+                gen.writeString(content.toString());
+                break;
+
+            default:
+                throw LOGGER.logExceptionAsError(
+                    new IllegalStateException("Unsupported BinaryData content type: " + content.getClass().getName()));
         }
     }
 }
