+++ sdk/core/azure-core/src/main/java/com/azure/core/util/binarydata/BinaryDataContentType.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.core.implementation.util;
+package com.azure.core.util.binarydata;
 
 import com.azure.core.util.BinaryData;
