+++ sdk/core/azure-core/src/main/java/com/azure/core/util/binarydata/ByteBufferContent.java
@@ -1,43 +1,36 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.core.implementation.util;
+package com.azure.core.util.binarydata;
 
 import com.azure.core.implementation.ImplUtils;
 import com.azure.core.util.FluxUtil;
-import com.azure.core.util.logging.ClientLogger;
 import com.azure.core.util.serializer.ObjectSerializer;
 import com.azure.core.util.serializer.TypeReference;
+import com.azure.json.JsonWriter;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
+import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
-import java.nio.Buffer;
 import java.nio.ByteBuffer;
 import java.nio.channels.AsynchronousByteChannel;
 import java.nio.channels.WritableByteChannel;
 import java.nio.charset.StandardCharsets;
-import java.util.List;
 import java.util.Objects;
 import java.util.concurrent.atomic.AtomicReferenceFieldUpdater;
 
-import static com.azure.core.util.FluxUtil.monoError;
-
 /**
- * A {@link BinaryDataContent} implementation which is backed by a {@link List} of {@link ByteBuffer}.
+ * A {@link BinaryDataContent} implementation which is backed by a {@link ByteBuffer}.
  */
-public class ListByteBufferContent extends BinaryDataContent {
-    private static final ClientLogger LOGGER = new ClientLogger(ListByteBufferContent.class);
-
-    private final List<ByteBuffer> content;
+public final class ByteBufferContent extends BinaryDataContent {
+    private final ByteBuffer content;
 
     private volatile byte[] bytes;
-    private static final AtomicReferenceFieldUpdater<ListByteBufferContent, byte[]> BYTES_UPDATER
-        = AtomicReferenceFieldUpdater.newUpdater(ListByteBufferContent.class, byte[].class, "bytes");
-
-    private Long cachedLength;
+    private static final AtomicReferenceFieldUpdater<ByteBufferContent, byte[]> BYTES_UPDATER
+        = AtomicReferenceFieldUpdater.newUpdater(ByteBufferContent.class, byte[].class, "bytes");
 
     /**
      * Creates a new instance of {@link BinaryDataContent}.
@@ -45,16 +38,13 @@ public class ListByteBufferContent extends BinaryDataContent {
      * @param content The {@link ByteBuffer} content.
      * @throws NullPointerException If {@code content} is null.
      */
-    public ListByteBufferContent(List<ByteBuffer> content) {
+    public ByteBufferContent(ByteBuffer content) {
         this.content = Objects.requireNonNull(content, "'content' cannot be null.");
     }
 
     @Override
     public Long getLength() {
-        if (cachedLength == null) {
-            cachedLength = content.stream().mapToLong(Buffer::remaining).sum();
-        }
-        return cachedLength;
+        return (long) content.remaining();
     }
 
     @Override
@@ -74,40 +64,47 @@ public class ListByteBufferContent extends BinaryDataContent {
 
     @Override
     public InputStream toStream() {
-        return new IterableOfByteBuffersInputStream(content);
+        return new ByteArrayInputStream(toBytes());
     }
 
     @Override
     public ByteBuffer toByteBuffer() {
-        return ByteBuffer.wrap(toBytes()).asReadOnlyBuffer();
+        return content.asReadOnlyBuffer();
     }
 
     @Override
     public Flux<ByteBuffer> toFluxByteBuffer() {
-        return Flux.fromIterable(content).map(ByteBuffer::asReadOnlyBuffer);
+        return Flux.just(content).map(ByteBuffer::asReadOnlyBuffer);
     }
 
     @Override
     public void writeTo(OutputStream outputStream) throws IOException {
-        for (ByteBuffer bb : content) {
-            ImplUtils.writeByteBufferToStream(bb.asReadOnlyBuffer(), outputStream);
-        }
+        Objects.requireNonNull(outputStream, "'outputStream' cannot be null");
+
+        ImplUtils.writeByteBufferToStream(toByteBuffer(), outputStream);
     }
 
     @Override
     public void writeTo(WritableByteChannel channel) throws IOException {
-        for (ByteBuffer bb : content) {
-            ImplUtils.fullyWriteBuffer(bb.asReadOnlyBuffer(), channel);
-        }
+        Objects.requireNonNull(channel, "'channel' cannot be null");
+
+        ImplUtils.fullyWriteBuffer(toByteBuffer(), channel);
     }
 
     @Override
     public Mono<Void> writeTo(AsynchronousByteChannel channel) {
         if (channel == null) {
-            return monoError(LOGGER, new NullPointerException("'channel' cannot be null."));
+            return Mono.error(new NullPointerException("'channel' cannot be null"));
         }
 
         return FluxUtil.writeToAsynchronousByteChannel(toFluxByteBuffer(), channel);
+    }
+
+    @Override
+    public void writeTo(JsonWriter jsonWriter) throws IOException {
+        Objects.requireNonNull(jsonWriter, "'jsonWriter' cannot be null.");
+
+        jsonWriter.writeBinary(toBytes());
     }
 
     @Override
@@ -131,20 +128,11 @@ public class ListByteBufferContent extends BinaryDataContent {
     }
 
     private byte[] getBytes() {
-        long length = getLength();
-        if (length > MAX_ARRAY_SIZE) {
-            throw LOGGER.logExceptionAsError(new IllegalStateException(TOO_LARGE_FOR_BYTE_ARRAY + length));
-        }
+        byte[] bytes = new byte[content.remaining()];
 
-        byte[] bytes = new byte[(int) length];
-        int offset = 0;
-
-        for (ByteBuffer bb : content) {
-            bb = bb.duplicate();
-            int count = bb.remaining();
-            bb.get(bytes, offset, count);
-            offset += count;
-        }
+        content.mark();
+        content.get(bytes);
+        content.flip();
 
         return bytes;
     }
