+++ sdk/core/azure-core/src/main/java/com/azure/core/util/binarydata/SerializableContent.java
@@ -1,12 +1,14 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.core.implementation.util;
+package com.azure.core.util.binarydata;
 
 import com.azure.core.implementation.ImplUtils;
 import com.azure.core.util.FluxUtil;
+import com.azure.core.util.logging.ClientLogger;
 import com.azure.core.util.serializer.ObjectSerializer;
 import com.azure.core.util.serializer.TypeReference;
+import com.azure.json.JsonWriter;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
@@ -21,29 +23,35 @@ import java.nio.charset.StandardCharsets;
 import java.util.Objects;
 import java.util.concurrent.atomic.AtomicReferenceFieldUpdater;
 
+import static com.azure.core.util.FluxUtil.monoError;
+
 /**
- * A {@link BinaryDataContent} implementation which is backed by a {@link ByteBuffer}.
+ * A {@link BinaryDataContent} implementation which is backed by a serializable object.
  */
-public final class ByteBufferContent extends BinaryDataContent {
-    private final ByteBuffer content;
+public final class SerializableContent extends BinaryDataContent {
+    private static final ClientLogger LOGGER = new ClientLogger(SerializableContent.class);
+
+    private final Object content;
+    private final ObjectSerializer serializer;
 
     private volatile byte[] bytes;
-    private static final AtomicReferenceFieldUpdater<ByteBufferContent, byte[]> BYTES_UPDATER
-        = AtomicReferenceFieldUpdater.newUpdater(ByteBufferContent.class, byte[].class, "bytes");
+    private static final AtomicReferenceFieldUpdater<SerializableContent, byte[]> BYTES_UPDATER
+        = AtomicReferenceFieldUpdater.newUpdater(SerializableContent.class, byte[].class, "bytes");
 
     /**
-     * Creates a new instance of {@link BinaryDataContent}.
-     *
-     * @param content The {@link ByteBuffer} content.
-     * @throws NullPointerException If {@code content} is null.
+     * Creates a new instance of {@link SerializableContent}.
+     * @param content The serializable object that forms the content of this instance.
+     * @param serializer The serializer that serializes the {@code content}.
+     * @throws NullPointerException if {@code serializer} is null.
      */
-    public ByteBufferContent(ByteBuffer content) {
-        this.content = Objects.requireNonNull(content, "'content' cannot be null.");
+    public SerializableContent(Object content, ObjectSerializer serializer) {
+        this.content = content;
+        this.serializer = Objects.requireNonNull(serializer, "'serializer' cannot be null.");
     }
 
     @Override
     public Long getLength() {
-        return (long) content.remaining();
+        return this.content == null ? null : (long) toBytes().length;
     }
 
     @Override
@@ -63,24 +71,24 @@ public final class ByteBufferContent extends BinaryDataContent {
 
     @Override
     public InputStream toStream() {
-        return new ByteArrayInputStream(toBytes());
+        return new ByteArrayInputStream(getBytes());
     }
 
     @Override
     public ByteBuffer toByteBuffer() {
-        return content.asReadOnlyBuffer();
+        return ByteBuffer.wrap(toBytes()).asReadOnlyBuffer();
     }
 
     @Override
     public Flux<ByteBuffer> toFluxByteBuffer() {
-        return Flux.just(content).map(ByteBuffer::asReadOnlyBuffer);
+        return Mono.fromSupplier(() -> ByteBuffer.wrap(toBytes())).flux();
     }
 
     @Override
     public void writeTo(OutputStream outputStream) throws IOException {
         Objects.requireNonNull(outputStream, "'outputStream' cannot be null");
 
-        ImplUtils.writeByteBufferToStream(toByteBuffer(), outputStream);
+        outputStream.write(toBytes());
     }
 
     @Override
@@ -93,10 +101,17 @@ public final class ByteBufferContent extends BinaryDataContent {
     @Override
     public Mono<Void> writeTo(AsynchronousByteChannel channel) {
         if (channel == null) {
-            return Mono.error(new NullPointerException("'channel' cannot be null"));
+            return monoError(LOGGER, new NullPointerException("'channel' cannot be null"));
         }
 
         return FluxUtil.writeToAsynchronousByteChannel(toFluxByteBuffer(), channel);
+    }
+
+    @Override
+    public void writeTo(JsonWriter jsonWriter) throws IOException {
+        Objects.requireNonNull(jsonWriter, "'jsonWriter' cannot be null.");
+
+        jsonWriter.writeRawValue(toString());
     }
 
     @Override
@@ -116,16 +131,10 @@ public final class ByteBufferContent extends BinaryDataContent {
 
     @Override
     public BinaryDataContentType getContentType() {
-        return BinaryDataContentType.BINARY;
+        return BinaryDataContentType.OBJECT;
     }
 
     private byte[] getBytes() {
-        byte[] bytes = new byte[content.remaining()];
-
-        content.mark();
-        content.get(bytes);
-        content.flip();
-
-        return bytes;
+        return serializer.serializeToBytes(content);
     }
 }
