+++ sdk/core/azure-core/src/main/java/com/azure/core/util/binarydata/StringContent.java
@@ -1,13 +1,14 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.core.implementation.util;
+package com.azure.core.util.binarydata;
 
 import com.azure.core.implementation.ImplUtils;
 import com.azure.core.util.FluxUtil;
 import com.azure.core.util.logging.ClientLogger;
 import com.azure.core.util.serializer.ObjectSerializer;
 import com.azure.core.util.serializer.TypeReference;
+import com.azure.json.JsonWriter;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
@@ -25,37 +26,34 @@ import java.util.concurrent.atomic.AtomicReferenceFieldUpdater;
 import static com.azure.core.util.FluxUtil.monoError;
 
 /**
- * A {@link BinaryDataContent} implementation which is backed by a serializable object.
+ * A {@link BinaryDataContent} implementation which is backed by a {@code String}.
  */
-public final class SerializableContent extends BinaryDataContent {
-    private static final ClientLogger LOGGER = new ClientLogger(SerializableContent.class);
+public final class StringContent extends BinaryDataContent {
+    private static final ClientLogger LOGGER = new ClientLogger(StringContent.class);
 
-    private final Object content;
-    private final ObjectSerializer serializer;
+    private final String content;
 
     private volatile byte[] bytes;
-    private static final AtomicReferenceFieldUpdater<SerializableContent, byte[]> BYTES_UPDATER
-        = AtomicReferenceFieldUpdater.newUpdater(SerializableContent.class, byte[].class, "bytes");
+    private static final AtomicReferenceFieldUpdater<StringContent, byte[]> BYTES_UPDATER
+        = AtomicReferenceFieldUpdater.newUpdater(StringContent.class, byte[].class, "bytes");
 
     /**
-     * Creates a new instance of {@link SerializableContent}.
-     * @param content The serializable object that forms the content of this instance.
-     * @param serializer The serializer that serializes the {@code content}.
-     * @throws NullPointerException if {@code serializer} is null.
+     * Creates a new instance of {@link StringContent}.
+     * @param content The string content.
+     * @throws NullPointerException if {@code content} is null.
      */
-    public SerializableContent(Object content, ObjectSerializer serializer) {
-        this.content = content;
-        this.serializer = Objects.requireNonNull(serializer, "'serializer' cannot be null.");
+    public StringContent(String content) {
+        this.content = Objects.requireNonNull(content, "'content' cannot be null.");
     }
 
     @Override
     public Long getLength() {
-        return this.content == null ? null : (long) toBytes().length;
+        return (long) toBytes().length;
     }
 
     @Override
     public String toString() {
-        return new String(toBytes(), StandardCharsets.UTF_8);
+        return this.content;
     }
 
     @Override
@@ -70,7 +68,7 @@ public final class SerializableContent extends BinaryDataContent {
 
     @Override
     public InputStream toStream() {
-        return new ByteArrayInputStream(getBytes());
+        return new ByteArrayInputStream(toBytes());
     }
 
     @Override
@@ -107,6 +105,13 @@ public final class SerializableContent extends BinaryDataContent {
     }
 
     @Override
+    public void writeTo(JsonWriter jsonWriter) throws IOException {
+        Objects.requireNonNull(jsonWriter, "'jsonWriter' cannot be null.");
+
+        jsonWriter.writeString(content);
+    }
+
+    @Override
     public boolean isReplayable() {
         return true;
     }
@@ -123,10 +128,10 @@ public final class SerializableContent extends BinaryDataContent {
 
     @Override
     public BinaryDataContentType getContentType() {
-        return BinaryDataContentType.OBJECT;
+        return BinaryDataContentType.TEXT;
     }
 
     private byte[] getBytes() {
-        return serializer.serializeToBytes(content);
+        return this.content.getBytes(StandardCharsets.UTF_8);
     }
 }
