+++ sdk/core/azure-core/src/main/java/com/azure/core/util/binarydata/package-info.java
@@ -1,0 +1,8 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+/**
+ * This package contains classes related to {@link com.azure.core.util.BinaryData}, specifically
+ * {@link com.azure.core.util.binarydata.BinaryDataContent} and its subtypes.
+ */
+package com.azure.core.util.binarydata;
