+++ sdk/core/azure-core/src/test/java/com/azure/core/http/rest/RestProxyTests.java
@@ -20,10 +20,9 @@ import com.azure.core.http.HttpRequest;
 import com.azure.core.http.HttpResponse;
 import com.azure.core.http.MockHttpResponse;
 import com.azure.core.implementation.http.rest.RestProxyUtils;
-import com.azure.core.implementation.util.BinaryDataContent;
-import com.azure.core.implementation.util.BinaryDataHelper;
 import com.azure.core.util.BinaryData;
 import com.azure.core.util.Context;
+import com.azure.core.util.binarydata.BinaryDataContent;
 import org.junit.jupiter.api.Named;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -161,14 +160,14 @@ public class RestProxyTests {
     public void doesNotChangeBinaryDataContentType(BinaryData data, long contentLength) {
         LocalHttpClient client = new LocalHttpClient();
         HttpPipeline pipeline = new HttpPipelineBuilder().httpClient(client).build();
-        Class<? extends BinaryDataContent> expectedContentClazz = BinaryDataHelper.getContent(data).getClass();
+        Class<? extends BinaryDataContent> expectedContentClazz = data.getContent().getClass();
 
         TestInterface testInterface = RestProxy.create(TestInterface.class, pipeline);
         Response<Void> response = testInterface.testMethod(data, "application/json", contentLength).block();
         assertEquals(200, response.getStatusCode());
 
         Class<? extends BinaryDataContent> actualContentClazz
-            = BinaryDataHelper.getContent(client.getLastHttpRequest().getBodyAsBinaryData()).getClass();
+            = client.getLastHttpRequest().getBodyAsBinaryData().getContent().getClass();
         assertEquals(expectedContentClazz, actualContentClazz);
     }
