+++ sdk/core/azure-core/src/test/java/com/azure/core/util/binarydata/MyFileContent.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.core.implementation.util;
+package com.azure.core.util.binarydata;
 
 import java.io.IOException;
 import java.nio.channels.AsynchronousFileChannel;
